requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-formtools==2.2
django-import-export==2.4.0
django-reversion==3.0.8
django-simpleui==2021.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
face-recognition-models==0.3.0
future==0.18.2
httplib2==0.9.2
ipython==7.16.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
joblib==1.0.0
MarkupPy==1.14
mysqlclient==2.0.2
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.5
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
psutil==5.8.0
Pygments==2.7.3
PyJWT==1.7.1
pytz==2020.5
PyYAML==5.3.1
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tablib==3.0.0
threadpoolctl==2.1.0
traitlets==4.3.3
wcwidth==0.2.5
wincertstore==0.2
xadmin @ git+git://github.com/sshwsfc/xadmin.git@47481fe8d784f4d6b2900c637467e3ce043fc575
xlrd==2.0.1
xlwt==1.3.0
